Make the build process more repeatable
This commit modifies the manpage generation process to parse the release
date out of `VERSION.dat`, instead of using the date/time when `pod2man`
was executed.

Embedding the date for the release into the generated manpages helps
ensure that the build process as a whole is more repeatable.

CLA: trivial

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>